requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
mediapipe>=0.10.0
